added two files of an exercise
I added an html with a solution of the "FizzFuzz" exercise. You can imput a number and see  the results in the console.  this is not for the previous page, but still is an exercise on the odinProject page so i added here. I go to  be doing this for rest of exercises.